backend/requirements.txt
fastapi
uvicorn
python-dotenv
groq
faiss-cpu
sentence-transformers
pymupdf
python-docx
python-pptx
requests
python-multipart
aiofiles
